(#893) Make LevelEditor transparent
1. Move the definition of struct LevelEditor to header
2. Store the source code of the Supa Script in the struct LevelEditor
